+++ cmd/milvus/boring_disabled.go
@@ -2,8 +2,12 @@
 
 package milvus
 
+import "github.com/milvus-io/milvus/pkg/v3/util/fips"
+
 func boringEnabled() bool {
 	return false
 }
 
-func maybeEnableOpenSSLFIPS() {}
+func maybeEnableOpenSSLFIPS() bool {
+	return fips.MaybeEnableOpenSSLFIPS()
+}
